multipage-auth-dash-app/pyproject.toml
[project]
name = "dash-app"
version = "0.1.0"
description = "An example mutlipage dash plotly app with auth"
readme = "README.md"
authors = [
    { name = "Nic Whippey", email = "n.whippey@exeter.ac.uk" }
]
requires-python = ">=3.10,<3.12"
dependencies = [
    "dash>=2.18.2",
    "dash-bootstrap-components>=1.6.0",
    "imageio>=2.27.0",
    "numpy>=1.26.4",
    "scikit-learn>=1.5.1",
    "flask-session>=0.8.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-login>=0.6.3",
]

[dependency-groups]
dev = [
    "pytest-mock>=3.14.0",
    "pytest>=8.3.4",
    "ruff>=0.8.1",
]

[project.scripts]
dash-app = "main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.pytest.ini_options]
tmp_path_retention_policy = "none"
